# ✨ Feat: (새로운 기능 추가) # 🐛 Fix: (버그 수정) # 📝 Docs: (문서 수정) # 💄 Style: (코드 스타일 변경) ♻️ Refactor: gateway와 auth로 나눠 코드 위치 리팩터링 (아직 미완성) # 🚨 Test: (테스트 코드 추가 / 수정) # 🔧 Chore: (빌드 업무 수정) # ⚡️ Perf: (성능 개선) # 🤖 CI: (CI/CD 설정 변경) # ⏪ Revert: (이전 커밋 되돌리기)
