+++ module-auth/src/main/java/com/example/moduleauth/config/PasswordEncoder.java
@@ -1,0 +1,17 @@
+package com.example.moduleauth.config;
+
+import at.favre.lib.crypto.bcrypt.BCrypt;
+import org.springframework.stereotype.Component;
+
+@Component
+public class PasswordEncoder {
+	
+	public String encode(String rawPassword) {
+		return BCrypt.withDefaults().hashToString(BCrypt.MIN_COST, rawPassword.toCharArray());
+	}
+	
+	public boolean matches(String rawPassword, String encodedPassword) {
+		BCrypt.Result result = BCrypt.verifyer().verify(rawPassword.toCharArray(), encodedPassword);
+		return result.verified;
+	}
+}
