+++ module-auth/src/main/java/com/example/moduleauth/domain/auth/controller/AuthController.java
@@ -1,0 +1,31 @@
+package com.example.moduleauth.domain.auth.controller;
+
+import com.example.moduleauth.domain.auth.dto.request.LoginRequest;
+import com.example.moduleauth.domain.auth.dto.request.SignupRequest;
+import com.example.moduleauth.domain.auth.dto.response.AuthResponse;
+import com.example.moduleauth.domain.auth.service.AuthService;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/api")
+public class AuthController {
+	
+	private final AuthService authService;
+	
+	@PostMapping("/v1/auth/signup")
+	public ResponseEntity<AuthResponse> signup(@Valid @RequestBody SignupRequest request) {
+		return ResponseEntity.ok(authService.signup(request));
+	}
+	
+	@PostMapping("/v1/auth/login")
+	public ResponseEntity<AuthResponse> login(@Valid @RequestBody LoginRequest request) {
+		return ResponseEntity.ok(authService.login(request));
+	}
+}
