+++ module-auth/src/main/java/com/example/moduleauth/domain/auth/dto/request/LoginRequest.java
@@ -1,0 +1,14 @@
+package com.example.moduleauth.domain.auth.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.Getter;
+
+@Getter
+public class LoginRequest {
+	
+	@NotNull(message = "이메일을 입력해주세요.")
+	private String email;
+	
+	@NotNull(message = "비밀번호를 입력해주세요.")
+	private String password;
+}
