+++ module-auth/src/main/java/com/example/moduleauth/domain/auth/dto/response/AuthResponse.java
@@ -1,0 +1,11 @@
+package com.example.moduleauth.domain.auth.dto.response;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class AuthResponse {
+	
+	private final String accessToken;
+}
