+++ module-auth/src/main/java/com/example/moduleauth/domain/member/dto/request/DeleteMemberRequest.java
@@ -1,0 +1,11 @@
+package com.example.moduleauth.domain.member.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.Getter;
+
+@Getter
+public class DeleteMemberRequest {
+	
+	@NotNull(message = "비밀번호를 입력해주세요.")
+	private String password;
+}
