--- module-auth/src/main/java/com/example/moduleauth/util/JwtAuthenticationToken.java
@@ -1,21 +1,0 @@
-package com.example.moduleauth.util;
-
-import org.springframework.security.authentication.AbstractAuthenticationToken;
-import com.example.moduleauth.entity.Auth;
-
-public class JwtAuthenticationToken extends AbstractAuthenticationToken {
-	
-	private final Auth auth;
-	
-	public JwtAuthenticationToken(Auth auth) {
-		super(auth.getAuthority());
-		this.auth = auth;
-		setAuthenticated(true);
-	}
-	
-	@Override
-	public Object getCredentials() { return null; }
-	
-	@Override
-	public Object getPrincipal() { return auth; }
-}
