+++ module-gateway/build.gradle
@@ -19,13 +19,16 @@ repositories {
 
 dependencies {
     implementation project(':module-common')
-    implementation project(':module-gateway')
-    implementation project(':module-ticket')
-    
-    implementation 'at.favre.lib:bcrypt:0.10.2'
+    implementation 'org.springframework.boot:spring-boot-starter-security'
+
+    // JWT
+    compileOnly group: 'io.jsonwebtoken', name: 'jjwt-api', version: '0.11.5'
+    runtimeOnly group: 'io.jsonwebtoken', name: 'jjwt-impl', version: '0.11.5'
+    runtimeOnly group: 'io.jsonwebtoken', name: 'jjwt-jackson', version: '0.11.5'
 }
 
 tasks.named('test') {
     useJUnitPlatform()
 }
+
 tasks.register("prepareKotlinBuildScriptModel"){}
