+++ module-gateway/src/main/java/com/example/modulegateway/ModuleGatewayApplication.java
@@ -1,0 +1,13 @@
+package com.example.modulegateway;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ModuleGatewayApplication {
+	
+	public static void main(String[] args) {
+		SpringApplication.run(ModuleGatewayApplication.class, args);
+	}
+	
+}
