+++ module-gateway/src/main/java/com/example/modulegateway/config/SecurityConfig.java
@@ -1,0 +1,53 @@
+package com.example.modulegateway.config;
+
+import com.example.modulegateway.filter.JwtAuthenticationFilter;
+import com.example.modulegateway.role.MemberRole;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.servletapi.SecurityContextHolderAwareRequestFilter;
+
+@Configuration
+@EnableWebSecurity
+@RequiredArgsConstructor
+@EnableMethodSecurity
+public class SecurityConfig {
+	
+	private final JwtAuthenticationFilter authenticationFilter;
+	
+	@Bean
+	public PasswordEncoder passwordEncoder() { return new BCryptPasswordEncoder(); }
+
+	@Bean
+	public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+		http.csrf(AbstractHttpConfigurer::disable)
+			.sessionManagement(session -> session
+				.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+			.addFilterBefore(authenticationFilter, SecurityContextHolderAwareRequestFilter.class)
+			.formLogin(AbstractHttpConfigurer::disable)
+			.httpBasic(AbstractHttpConfigurer::disable)
+			.logout(AbstractHttpConfigurer::disable)
+			.rememberMe(AbstractHttpConfigurer::disable)
+			.authorizeHttpRequests(auth -> auth
+				.requestMatchers("/login.html").permitAll()
+				.requestMatchers("/point-charge.html").permitAll()
+				.requestMatchers("/actuator/**").permitAll()
+				.requestMatchers("/api/v1/auth/**").permitAll()
+				.requestMatchers(HttpMethod.GET,"/api/v1/games/**").authenticated()
+				.requestMatchers("/api/v1/admin/**").hasAuthority(MemberRole.Authority.ADMIN)
+				.requestMatchers("/actuator/prometheus").permitAll()
+				.anyRequest().authenticated()
+			);
+		
+		return http.build();
+	}
+}
