+++ module-gateway/src/main/java/com/example/modulegateway/entity/Auth.java
@@ -1,0 +1,23 @@
+package com.example.modulegateway.entity;
+
+import com.example.modulegateway.role.MemberRole;
+import java.util.List;
+import lombok.Getter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+
+@Getter
+public class Auth {
+	
+	private final Long id;
+	private final String email;
+	private final MemberRole role;
+	private final List<? extends GrantedAuthority> authority;
+	
+	public Auth(Long id, String email, MemberRole role) {
+		this.id = id;
+		this.email = email;
+		this.role = role;
+		this.authority = List.of(new SimpleGrantedAuthority(role.name()));
+	}
+}
