+++ module-gateway/src/main/resources/application.yml
@@ -1,0 +1,5 @@
+spring:
+  profiles:
+    active: local
+  application:
+    name: module-gateway
