+++ module-gateway/src/test/java/com/example/modulegateway/ModuleGatewayApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.modulegateway;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ModuleGatewayApplicationTests {
+	
+	@Test
+	void contextLoads() {
+	}
+	
+}
