+++ settings.gradle
@@ -3,3 +3,4 @@ include ':module-common'
 include ':module-ticket'
 include ':module-auth'
 include ':module-game'
+include ':module-gateway'
